Add Best of 5 rounds play mode with init. prompt.
Added a 5 rounds function and an initial prompt that asks the players if
they want to play or not. Also, made a condition if the player input
isn't one of the 3 options (case insensitive for rock, paper, scissors)
the round will be replayed without awarding points to anyone.

So far the game only receives input from prompts and display the results
on the console. Hopefully, later on with help of other projects, a GUI
will be added.