Replaces prompts with buttons + GUI
Changed the program from being played with prompts and displaying
results on the console to 3 buttons for the users to make their play
and a basic score display.
Added a function that prevents the game from accumulating scores after
either player reaches 5 wins.
Also, added a Start Over button that resets the game values and lets
the user play again.